// lib/shopify/types.ts
export type Maybe<T> = T | null;

export type Connection<T> = {
    edges: Array<Edge<T>>;
};

export type Edge<T> = {
    node: T;
};

export type Cart = {
    id: string;
    checkoutUrl: string;
    cost: {
        subtotalAmount: Money;
        totalAmount: Money;
        totalTaxAmount: Money;
    };
    lines: Connection<CartItem>;
    totalQuantity: number;
};

export type CartItem = {
    id: string;
    quantity: number;
    cost: {
        totalAmount: Money;
    };
    merchandise: {
        id: string;
        title: string;
        selectedOptions: {
            name: string;
            value: string;
        }[];
        product: Product;
    };
};

export type Collection = {
    id?: string;
    handle: string;
    title: string;
    description: string;
    seo: SEO;
    updatedAt: string;
    image?: Image;
    products: Connection<Product>;
};

export type PredictiveSearchResult = {
    products: Product[];
    collections: Collection[];
};

export type Image = {
    url: string;
    altText: string;
    width: number;
    height: number;
};

export type Menu = {
    title: string;
    path: string;
};

export type Money = {
    amount: string;
    currencyCode: string;
};

export type Page = {
    id: string;
    title: string;
    handle: string;
    body: string;
    bodySummary: string;
    seo?: SEO;
    createdAt: string;
    updatedAt: string;
};

export type Product = {
    id: string;
    handle: string;
    availableForSale: boolean;
    title: string;
    description: string;
    descriptionHtml: string;
    options: ProductOption[];
    priceRange: {
        maxVariantPrice: Money;
        minVariantPrice: Money;
    };
    variants: Connection<ProductVariant>;
    featuredImage: Image;
    images: Connection<Image>;
    seo: SEO;
    tags: string[];
    updatedAt: string;
    // Custom metafields for TAP
    material?: string;
    use_case?: string;
    durability?: string;
    reflective?: boolean;
    made_in?: string;
    cause?: string;
    preorder?: boolean;
    metafields?: {
        namespace: string;
        key: string;
        value: string;
    }[];
};

export type ProductOption = {
    id: string;
    name: string;
    values: string[];
};

export type ProductVariant = {
    id: string;
    title: string;
    availableForSale: boolean;
    selectedOptions: {
        name: string;
        value: string;
    }[];
    price: Money;
};

export type SEO = {
    title: string;
    description: string;
};

export type ShopifyExtension = {
    cost: {
        actualQueryCost: number;
        requestedQueryCost: number;
    };
};